refactor: move IOperations types the operation context for IOperations file
